+++ src/detectors/JavaScript/JavaScriptComponentDetector.ts
@@ -14,19 +14,7 @@ export class JavaScriptComponentDetector implements IProjectComponentDetector {
   }
 
   async detectComponent(langAtPath: LanguageAtPath): Promise<ProjectComponent[]> {
-    const backendPackages = [
-      'express',
-      'node',
-      'koa',
-      'hapi',
-      'flatiron',
-      'locomotive',
-      'nodal',
-      '@adonisjs/framework',
-      'thinkjs',
-      'sails',
-      '@types/node',
-    ];
+    const backendPackages = ['express', 'koa', 'hapi', 'flatiron', 'locomotive', 'nodal', '@adonisjs/framework', 'thinkjs', 'sails'];
 
     const frontendPackages = ['webpack', 'jquery', 'gulp', 'grunt', 'browserify', 'babel'];