+++ src/practices/JavaScript/DeprecatedTSLintPractice.ts
@@ -23,6 +23,8 @@ export class DeprecatedTSLintPractice implements IPractice {
       if (ctx.packageInspector.hasPackage('tslint')) {
         return PracticeEvaluationResult.notPracticing;
       }
+
+      return PracticeEvaluationResult.practicing;
     }
 
     return PracticeEvaluationResult.unknown;