+++ src/practices/index.ts
@@ -1,9 +1,7 @@
-import { PullRequestPractice } from './Collaboration/PullRequestPractice';
 import { TypeScriptUsedPractice } from './JavaScript/TypeScriptUsedPractice';
 import { PrettierUsedPractice } from './JavaScript/PrettierUsedPractice';
 import { ESLintUsedPractice } from './JavaScript/ESLintUsedPractice';
 import { LockfileIsPresentPractice } from './PackageManagement/LockfileIsPresentPractice';
-import { UnitTestPractice } from './Testing/UnitTestPractice';
 import { JsFrontendTestingFrameworkUsedPractice } from './JavaScript/JsFrontendTestingFrameworkUsedPractice';
 import { JsLoggerUsedPractice } from './JavaScript/JsLoggerUsedPractice';
 import { LicenseIsPresentPractice } from './LanguageIndependent/LicenseIsPresentPractice';
@@ -18,18 +16,16 @@ import { DockerizationUsedPractice } from './LanguageIndependent/DockerizationUs
 import { EditorConfigIsPresentPractice } from './LanguageIndependent/EditorConfigIsPresentPractice';
 import { JsGitignoreIsPresentPractice } from './JavaScript/JsGitignoreIsPresentPractice';
 import { JsGitignoreCorrectlySetPractice } from './JavaScript/JsGitignoreCorrectlySetPractice';
-import { DependenciesVersionPractice } from './LanguageIndependent/DependenciesVersionPractice';
+import { DependenciesVersionPractice } from './JavaScript/DependenciesVersionPractice';
 import { ESLintWithoutErrorsPractice } from './JavaScript/ESLintWithoutErrorsPractice';
 
 // register practices here
 export const practices = [
-  PullRequestPractice,
   TypeScriptUsedPractice,
   PrettierUsedPractice,
   ESLintUsedPractice,
   ESLintWithoutErrorsPractice,
   LockfileIsPresentPractice,
-  UnitTestPractice,
   JsFrontendTestingFrameworkUsedPractice,
   JsBackendTestingFrameworkUsedPractice,
   JsLoggerUsedPractice,